fix: restore original Xiaohongshu sync flow
