+++ packages/extension/__tests__/xiaohongshu-prosemirror.test.ts
@@ -2,11 +2,8 @@ import { describe, expect, it } from 'vitest'
 import { markdownToXiaohongshuProseMirror } from '../src/adapters/xiaohongshu-prosemirror'
 
 describe('markdownToXiaohongshuProseMirror', () => {
-  it('converts paragraphs without leaving raw line breaks in text nodes', () => {
-    const result = markdownToXiaohongshuProseMirror(
-      '第一行\n第二行\n\n第三段',
-      new Map()
-    )
+  it('converts paragraphs without leaving raw line breaks in text nodes', async () => {
+    const result = await markdownToXiaohongshuProseMirror('第一行\n第二行\n\n第三段')
 
     expect(result).toEqual({
       type: 'doc',
@@ -26,18 +23,14 @@ describe('markdownToXiaohongshuProseMirror', () => {
     })
   })
 
-  it('matches Xiaohongshu heading, marks, list and image node names', () => {
-    const images = new Map([
-      ['https://example.com/a.png', {
+  it('matches Xiaohongshu heading, marks, list and image node names', async () => {
+    const result = await markdownToXiaohongshuProseMirror(
+      '# 标题\n\n**粗体**和*斜体*\n\n- 列表\n\n![图片](https://example.com/a.png)',
+      async () => ({
         url: 'https://ros-preview.xhscdn.com/file-id',
         width: 800,
         height: 600,
-      }],
-    ])
-
-    const result = markdownToXiaohongshuProseMirror(
-      '# 标题\n\n**粗体**和*斜体*\n\n- 列表\n\n![图片](https://example.com/a.png)',
-      images
+      })
     )
 
     expect(result.content?.map(node => node.type)).toEqual([
@@ -47,9 +40,9 @@ describe('markdownToXiaohongshuProseMirror', () => {
       'image',
     ])
     expect(result.content?.[1].content).toEqual([
-      { type: 'text', text: '粗体', marks: [{ type: 'bold' }] },
+      { type: 'text', text: '粗体', marks: [{ type: 'highlight' }] },
       { type: 'text', text: '和' },
-      { type: 'text', text: '斜体', marks: [{ type: 'italic' }] },
+      { type: 'text', text: '斜体' },
     ])
     expect(result.content?.[3]).toEqual({
       type: 'image',
