+++ packages/extension/src/adapters/index.ts
@@ -45,7 +45,7 @@ import {
   EastmoneyAdapter,
 } from '@wechatsync/core'
 import { XiaoheiheAdapter } from './xiaoheihe'
-import { XiaohongshuLongformAdapter } from './xiaohongshu-longform'
+import { XiaohongshuAdapter } from './xiaohongshu'
 
 // 私有适配器 - private/ 目录通过 git submodule 管理
 const privateModules = import.meta.glob<Record<string, unknown>>(
@@ -100,7 +100,7 @@ const ADAPTER_CLASSES: AdapterConstructor[] = [
   ZipDownloadAdapter,
   EastmoneyAdapter,
   XiaoheiheAdapter,
-  XiaohongshuLongformAdapter,
+  XiaohongshuAdapter,
   ...getPrivateAdapters(),
 ]
